requirements.txt
Flask==2.0.3
gunicorn==20.1.0
pandas==1.4.2
python-dotenv==0.20.0
python-telegram-bot==13.11
discord-webhook==0.15.0